feat: Prepare Mainnet Build of Android/iOS for Test Event (#51)
* chore: add ttd currency

* feat: ensure account initialization

* chore: set currency default on store init

* chore: pin skips phone number

* fix: remove rewards for test event

* chore: remove unnecessary notifications

* chore: bump to 1.34.2

* fix: reduce selected vendor sheet height

* chore: remove vendors from services

* feat: dynamic vendor links

* chore: enable firebase for alfajores builds

* chore: remove vendor from env file

* feat: prepare mainnet envs for deployment

* feat: add dynamic firebase rtdb url

* chore: bump build number

* fix: fixed android crashin on build (#44)

* feat: upgrade to api 31

* chore: update project pbx

* chore: set map def to trinidad

* feat: map ref custom hook

* chore: sort vendors reverse (zed on top)

* chore: pull in map hook usage

* chore: bump to (26)

* fix: android launch permissions

* chore: bump build 27

* fix: remove unused testnet banner

* fix: release build using debug jsbundle

* chore: add screenshots for android store

* fix: resize images

* chore: replace content text

* chore: bump version 30

* cicd: add script for deploying codepush

* cicd: remove unused emerge

* chore: update project pbx
