+++ .env.alfajores
@@ -9,7 +9,7 @@ DEFAULT_SYNC_MODE=5
 DEV_SETTINGS_ACTIVE_INITIALLY=false
 FIREBASE_ENABLED=true
 SHOW_TESTNET_BANNER=true
-SHOW_GET_INVITE_LINK=true
+SHOW_GET_INVITE_LINK=false
 APP_BUNDLE_ID=io.zed.alfajores.dev
 APP_STORE_ID=1482389446
 APP_DISPLAY_NAME=Alfajores
