+++ android/app/src/main/AndroidManifest.xml
@@ -12,7 +12,7 @@
   <uses-permission android:name="android.permission.CAMERA" />
 
   <application android:name=".MainApplication" android:allowBackup="false" android:icon="@mipmap/ic_launcher" android:roundIcon="@mipmap/ic_launcher_round" android:label="@string/app_name" android:theme="@style/AppTheme" android:networkSecurityConfig="@xml/network_security_config">
-    <activity android:name=".MainActivity" android:screenOrientation="portrait" android:configChanges="keyboard|keyboardHidden|orientation|screenSize|uiMode" android:label="@string/app_name" android:launchMode="singleTop" android:theme="@style/AppTheme" android:windowSoftInputMode="adjustResize">
+    <activity android:name=".MainActivity" android:exported="true" tools:node="merge" android:screenOrientation="portrait" android:configChanges="keyboard|keyboardHidden|orientation|screenSize|uiMode" android:label="@string/app_name" android:launchMode="singleTop" android:theme="@style/AppTheme" android:windowSoftInputMode="adjustResize">
       <intent-filter>
         <action android:name="android.intent.action.MAIN" />
         <category android:name="android.intent.category.LAUNCHER" />
@@ -47,13 +47,13 @@
 
     <meta-data android:name="CLEVERTAP_BACKGROUND_SYNC" android:value="1"/>
 
-    <service android:name=".FirebaseMessagingService">
+    <service android:name=".FirebaseMessagingService" android:exported="true">
       <intent-filter>
         <action android:name="com.google.firebase.MESSAGING_EVENT"/>
       </intent-filter>
     </service>
 
-    <service android:name="com.clevertap.android.sdk.pushnotification.CTNotificationIntentService" android:exported="false">
+    <service android:name="com.clevertap.android.sdk.pushnotification.CTNotificationIntentService" android:exported="true">
       <intent-filter>
         <action android:name="com.clevertap.PUSH_EVENT" />
       </intent-filter>
