+++ android/app/src/main/java/org/celo/mobile/MainApplication.java
@@ -30,7 +30,7 @@ public class MainApplication
 
     @Override
     protected String getJSBundleFile() {
-      return CodePush.getJSBundleFile("main.jsbundle");
+      return CodePush.getJSBundleFile();
     }
 
     @Override
