+++ fastlane/Fastfile
@@ -10,7 +10,7 @@ def fastlane_emerge(environment:, file_path:)
   # Note: this requires the EMERGE_API_TOKEN env var to be set
   emerge(
     file_path: file_path,
-    repo_name: "valora-inc/wallet",
+    repo_name: "zed-io/kolektivo-wallet",
     sha: `git rev-parse HEAD`.strip,
     branch: `git branch --show-current`.strip,
     build_type: environment,
@@ -50,16 +50,12 @@ platform :android do
       track_promote_to: track,
       package_name: get_app_bundle_id(env),
       mapping: lane_context[SharedValues::GRADLE_MAPPING_TXT_OUTPUT_PATH],
+      release_status: 'draft',
       skip_upload_apk: true,
       skip_upload_metadata: true,
       skip_upload_changelogs: true,
       skip_upload_images: true,
       skip_upload_screenshots: true
-    )
-
-    fastlane_emerge(
-      environment: env, 
-      file_path: lane_context[SharedValues::GRADLE_AAB_OUTPUT_PATH],
     )
   end
