+++ fastlane/metadata/android/en-US/title.txt
@@ -1,1 +1,1 @@
-Celo Alfajores Wallet
+Kolektivo
