--- fastlane/metadata/android/es-419/short_description.txt
@@ -1,1 +1,0 @@
-Monedero digital seguro que le permite enviar dinero a cualquier persona
