--- fastlane/metadata/android/es-419/title.txt
@@ -1,1 +1,0 @@
-Monedero Celo alfajores
