+++ src/map/constants.ts
@@ -3,13 +3,13 @@ import { LatLng, Region } from 'react-native-maps'
 export const BASE_TAG = 'MapScreen'
 
 export const LOCALE_LATLNG: LatLng = {
-  latitude: 12.1224,
-  longitude: -68.8824,
+  latitude: 10.6603,
+  longitude: -61.5086,
 }
 
 export const LOCALE_OFFSET: Omit<Region, 'longitude' | 'latitude'> = {
-  latitudeDelta: 0.1922,
-  longitudeDelta: 0.1421,
+  latitudeDelta: 0.0022,
+  longitudeDelta: 0.0421,
 }
 
 export const LOCALE_REGION: Region = { ...LOCALE_LATLNG, ...LOCALE_OFFSET }
