+++ src/vendors/utils.ts
@@ -1,9 +1,7 @@
 import BottomSheet from '@gorhom/bottom-sheet'
-import { map } from 'lodash'
-import React, { useEffect, useState } from 'react'
-import { LatLng } from 'react-native-maps'
+import { map, orderBy } from 'lodash'
+import React, { useEffect } from 'react'
 import { useSelector } from 'react-redux'
-import { LOCALE_LATLNG } from 'src/map/constants'
 import { currentVendorSelector } from 'src/vendors/selector'
 import { Vendor, Vendors, VendorWithLocation } from 'src/vendors/types'
 
@@ -16,7 +14,7 @@ export const formatVendors = (vendorObject: any): Vendors => {
   const { data } = vendorObject
   const result = Object.assign(
     {},
-    ...data.map((v: any) => {
+    ...orderBy(data, ['attributes.name'], ['desc']).map((v: any) => {
       const {
         name,
         subtitle,
@@ -58,23 +56,6 @@ export const hasValidLocation = (vendor: VendorWithLocation): boolean => {
   const { location } = vendor
   const { latitude, longitude } = location
   return !!latitude && !!longitude
-}
-
-/**
- * This custom hook will return the latest vendor selected from the
- * redux store as well as its computed location. If the location of
- * the latest vendor is invalid, it will return the latest location
- * from the local state of this hook.
- */
-export const useCurrentVendorLocation = () => {
-  const currentVendor = useSelector(currentVendorSelector) as VendorWithLocation
-  const [location, setLocation] = useState<LatLng>(LOCALE_LATLNG)
-  useEffect(() => {
-    if (currentVendor && hasValidLocation(currentVendor)) {
-      setLocation(currentVendor.location)
-    }
-  }, [currentVendor])
-  return { currentVendor, location }
 }
 
 export const useInteractiveBottomSheet = (bottomSheetRef: React.RefObject<BottomSheet>) => {
